+++ app/layout.tsx
@@ -1,6 +1,7 @@
 import { Metadata } from 'next';
 import Footer from '@/components/ui/Footer';
 import Navbar from '@/components/ui/Navbar';
+import TikiTaskbar from '@/components/ui/Taskbar/TikiTaskbar';
 import { Toaster } from '@/components/ui/Toasts/toaster';
 import Concierge from '@/components/ui/Agent/Concierge';
 import ClubAudio from '@/components/ui/Audio/ClubAudio';
@@ -73,6 +74,9 @@ export default async function RootLayout({ children }: PropsWithChildren) {
           id="skip"
           className="min-h-[calc(100dvh-4rem)] md:min-h[calc(100dvh-5rem)]"
         >
+          <Suspense>
+            <TikiTaskbar />
+          </Suspense>
           {children}
         </main>
         <Footer />